# docs/swagger/swagger.yaml
definitions:
...
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.LoginResult:
    properties:
      currentOrgId:
        type: string
      tokenPair:
        $ref: '#/definitions/jwt.TokenPair'
      user:
        $ref: '#/definitions/auth.User'
    type: object
  auth.LogoutRequest:
    properties:
      refresh_token:
        type: string
    type: object
  auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  auth.RefreshResponse:
    properties:
      tokenPair:
        $ref: '#/definitions/jwt.TokenPair'
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        maxLength: 32
        minLength: 8
        type: string
      phone:
        type: string
      username:
        maxLength: 32
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  auth.RegisterResponse:
    properties:
      email:
        type: string
      message:
        type: string
      role:
        type: string
      userid:
        type: string
      username:
        type: string
    type: object
  auth.User:
    properties:
      email:
        type: string
      id:
        type: string
      name:
        type: string
      role:
        type: string
    type: object
...
  jwt.TokenPair:
    properties:
      accessToken:
        type: string
      expiresIn:
        type: integer
      refreshToken:
        type: string
    type: object
...
  response.ErrResponse:
    properties:
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
  response.ErrorInfo:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
...
  response.SuccessResponse:
    properties:
      data: {}
      success:
        type: boolean
    type: object
...
host: api.lab.note
info:
  contact: {}
  description: |-
    CQRS-style API. У каждой команды/запроса — свой путь вида
    /api/commands/{name} или /api/queries/{name}, тело = data.
  title: Agro Platform API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email и пароль
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.LoginResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Вход
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh-токен
        in: body
        name: request
        schema:
          $ref: '#/definitions/auth.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Выход (текущая реализация — заглушка, blacklist токенов не ведётся)
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.RefreshResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Перевыпуск токена
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/auth.RegisterResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrResponse'
      summary: Регистрация
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: 'JWT из POST /auth/login. Формат: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"